+++ src/event/player/PlayerEnchantOptionsRequestEvent.php
@@ -21,13 +21,13 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\event\inventory;
+namespace pocketmine\event\player;
 
 use pocketmine\block\inventory\EnchantInventory;
 use pocketmine\event\Cancellable;
 use pocketmine\event\CancellableTrait;
 use pocketmine\event\Event;
-use pocketmine\item\enchantment\EnchantmentOption;
+use pocketmine\item\enchantment\EnchantOption;
 use pocketmine\player\Player;
 use pocketmine\utils\Utils;
 use function count;
@@ -36,21 +36,18 @@ use function count;
  * Called when a player inserts an item into an enchanting table's input slot.
  * The options provided by the event will be shown on the enchanting table menu.
  */
-class PlayerEnchantmentOptionsRequestEvent extends Event implements Cancellable{
+class PlayerEnchantOptionsRequestEvent extends PlayerEvent implements Cancellable{
 	use CancellableTrait;
 
 	/**
-	 * @param EnchantmentOption[] $options
+	 * @param EnchantOption[] $options
 	 */
 	public function __construct(
-		private readonly Player $player,
+		Player $player,
 		private readonly EnchantInventory $enchantInventory,
 		private array $options
 	){
-	}
-
-	public function getPlayer() : Player{
-		return $this->player;
+		$this->player = $player;
 	}
 
 	public function getEnchantInventory() : EnchantInventory{
@@ -58,17 +55,17 @@ class PlayerEnchantmentOptionsRequestEvent extends Event implements Cancellable{
 	}
 
 	/**
-	 * @return EnchantmentOption[]
+	 * @return EnchantOption[]
 	 */
 	public function getOptions() : array{
 		return $this->options;
 	}
 
 	/**
-	 * @param EnchantmentOption[] $options
+	 * @param EnchantOption[] $options
 	 */
 	public function setOptions(array $options) : void{
-		Utils::validateArrayValueType($options, function(EnchantmentOption $_) : void{ });
+		Utils::validateArrayValueType($options, function(EnchantOption $_) : void{ });
 		if(($optionCount = count($options)) > 3){
 			throw new \LogicException("The maximum number of options for an enchanting table is 3, but $optionCount have been passed");
 		}
